+++ rust/substreams/src/lib.rs
@@ -1,10 +1,14 @@
 mod externs;
+mod hex;
+
 pub mod log;
 pub mod memory;
 pub mod pb;
 pub mod proto;
 pub mod rpc;
 pub mod state;
+
+pub use crate::hex::Hex;
 
 pub fn output<M: prost::Message>(msg: M) {
     // Need to return the buffer and forget about it issue occured when trying to write large data